Update simulation unit tests
