+++ citam/engine/calculators/calculator.py
@@ -2,7 +2,7 @@ from citam.engine.core.agent import Agent
 from citam.engine.facility.indoor_facility import Facility
 
 from abc import ABC, abstractmethod
-from typing import List, Union
+from typing import List, OrderedDict, Union
 import os
 
 
@@ -13,11 +13,19 @@ class Calculator(ABC):
         self.kwargs = kwargs
 
     @abstractmethod
-    def run(self, current_step: int, agents: List[Agent]):
+    def initialize(
+        self,
+        agents: OrderedDict[int, Agent],
+        work_directory: Union[str, bytes, os.PathLike],
+    ):
         pass
 
     @abstractmethod
-    def save_to_files(
+    def run(self, current_step: int, agents: List[Agent], **kwargs):
+        pass
+
+    @abstractmethod
+    def finalize(
         self,
         agents: List[Agent],
         work_directory: Union[str, bytes, os.PathLike],
