+++ citam/engine/calculators/close_contacts_calculator.py
@@ -1,4 +1,6 @@
-from typing import List, Tuple, Union
+from typing import Dict, List, Optional, OrderedDict, TextIO, Tuple, Union
+
+from networkx.generators.random_graphs import random_shell_graph
 from citam.engine.calculators.calculator import Calculator
 from citam.engine.facility.indoor_facility import Facility
 from citam.engine.core.agent import Agent
@@ -8,6 +10,11 @@ import numpy as np
 import scipy.spatial.distance
 import os
 import json
+import xml.etree.ElementTree as ET
+
+
+ET.register_namespace("", "http://www.w3.org/2000/svg")
+ET.register_namespace("xlink", "http://www.w3.org/1999/xlink")
 
 
 class CloseContactsCalculator(Calculator):
@@ -16,12 +23,45 @@ class CloseContactsCalculator(Calculator):
         facility: Facility,
         contact_distance: float = 6.0,
     ) -> None:
-        super().__init__()
+        super().__init__(facility)
         self.facility = facility
-        self.contact_events = cev.ContactEvents()
         self.contact_distance = contact_distance
 
-    def run(self, current_step: int, agents: List[Agent]) -> None:
+        # Object to keep track of contact events
+        self.contact_events = cev.ContactEvents()
+
+        # list used to keep track of total contact events per xy location per
+        #  floor
+        self.step_contact_locations: List[Dict[Tuple[int, int], int]] = [
+            {} for _ in facility.floorplans
+        ]
+        self.data_outfiles = []
+
+    def initialize(
+        self, agents: OrderedDict[int, Agent], workdir: Union[os.PathLike, str]
+    ):
+        for a in agents.values():
+            # Make sure another calculator is not setting
+            #  the contact duration property
+            if "contact_duration" in a.cumulative_properties:
+                raise ValueError(
+                    "The close contacts calculator is clashing with another "
+                    "calculator. Please double check."
+                )
+            a.cumulative_properties["contact_duration"] = 0
+
+        # Opening output files
+        self.data_outfiles = []
+        for floor_number in range(self.facility.number_of_floors):
+            directory = os.path.join(workdir, "floor_" + str(floor_number))
+            contact_file = os.path.join(directory, "contacts.txt")
+            self.data_outfiles.append(open(contact_file, "w"))
+
+    def run(
+        self,
+        current_step: int,
+        active_agents: List[Agent],
+    ) -> None:
         """
         Iterate over agents, compute whether they fall within the contact
         distance or not and verify that they are indeed making contact (based
@@ -30,13 +70,28 @@ class CloseContactsCalculator(Calculator):
         :param agents: list of agents under consideration for contact
             statistics
         :type agents: List[Agent]
+        :param contact_outfiles: Files to write contact data, one per floor,
+                defaults to None
+        :type contact_outfiles: List[TextIO], optional
         """
         # Immediately discard contact event if either:
         # 1. agents are in different floors
         # 2. agents are in different spaces but only if the spaces
         #    are not neighbors in the hallway graphs
 
-        if not agents:
+        self.step_contact_locations = [
+            {}
+        ] * self.facility.number_of_floors  # One per floor
+
+        # At least 2 agents required.
+        if len(active_agents) <= 1:
+            return
+
+        # Only consider active agents
+        agents = [a for a in active_agents if a.current_location is not None]
+
+        if len(agents) == 0:
+            # Nothing to do
             return
 
         positions_vector = np.array([a.pos for a in agents])
@@ -63,6 +118,23 @@ class CloseContactsCalculator(Calculator):
                     self.verify_and_add_contact(
                         current_step, fn, agent1, agent2
                     )
+
+        for fn in range(self.facility.number_of_floors):
+
+            n_contact_loc = len(self.step_contact_locations[fn])
+            self.data_outfiles[fn].write(
+                str(n_contact_loc) + "\nstep :" + str(current_step) + "\n"
+            )
+
+            for cl, nc in self.step_contact_locations[fn].items():
+                self.data_outfiles[fn].write(
+                    str(cl[0]) + "\t" + str(cl[1]) + "\t" + str(nc) + "\n"
+                )
+
+    def finalize(self):
+
+        for cof in self.data_outfiles:
+            cof.close()
 
     def identify_xy_proximity(
         self, positions_vector: np.ndarray
@@ -132,8 +204,8 @@ class CloseContactsCalculator(Calculator):
         dx = (agent2.pos[0] - agent1.pos[0]) / 2.0
         dy = (agent2.pos[1] - agent1.pos[1]) / 2.0
         contact_pos = (agent1.pos[0] + dx, agent1.pos[1] + dy)
-        agent1.cumulative_contact_duration += 1
-        agent2.cumulative_contact_duration += 1
+        agent1.cumulative_properties["contact_duration"] += 1
+        agent2.cumulative_properties["contact_duration"] += 1
         self.contact_events.add_contact(
             agent1, agent2, current_step, contact_pos
         )
@@ -158,7 +230,7 @@ class CloseContactsCalculator(Calculator):
         agent_ids, n_contacts = [], []
         for unique_id, agent in agents.items():
             agent_ids.append("ID_" + str(unique_id + 1))
-            n_contacts.append(agent.cumulative_contact_duration)
+            n_contacts.append(agent.cumulative_properties["contact_duration"])
 
         return agent_ids, n_contacts
 
@@ -174,6 +246,8 @@ class CloseContactsCalculator(Calculator):
             saved.
         :type work_directory: str
         """
+
+        # TODO: Don't overwrite existing file
 
         # Total contacts per agent
         agent_ids, n_contacts = self.extract_contact_distribution_per_agent(
@@ -217,9 +291,7 @@ class CloseContactsCalculator(Calculator):
             )
 
             contacts_per_coord = (
-                self.contact_events.get_contacts_per_coordinates(
-                    self.current_step, floor_number
-                )
+                self.contact_events.get_contacts_per_coordinates(floor_number)
             )
 
             with open(coord_contact_dist_filename, "a") as outfile:
@@ -232,3 +304,62 @@ class CloseContactsCalculator(Calculator):
                     )
 
             self.create_svg_heatmap(contacts_per_coord, floor_directory)
+
+    def create_svg_heatmap(
+        self,
+        contacts_per_coord: Dict[Tuple[int, int], int],
+        floor_directory: str,
+    ) -> None:
+        """
+        Create and save a heatmap from coordinate contact data.
+
+        :param contacts_per_coord: dictionary where each key is an (x,y )
+                tuple and values are the number of contacts in that location.
+        :type contacts_per_coord: Dict[Tuple[int, int], int]
+        :param floor_directory: directory to save the heatmap file.
+        :type floor_directory: str
+        :raises FileNotFoundError: if the floor svg map file is not found.
+        """
+
+        heatmap_file = os.path.join(floor_directory, "heatmap.svg")
+        map_file = os.path.join(floor_directory, "map.svg")
+
+        if not os.path.isfile(map_file):
+            raise FileNotFoundError(map_file)
+
+        max_contacts = 0
+        if len(contacts_per_coord.values()) > 0:
+            max_contacts = max(contacts_per_coord.values())
+        tree = ET.parse(map_file)
+
+        root = tree.getroot()[0]
+
+        filter_elem = ET.Element("filter")
+        filter_elem.set("id", "blurMe")
+        filter_elem.set("width", str(self.contact_distance * 2))
+        filter_elem.set("height", str(self.contact_distance * 2))
+        filter_elem.set("x", "-100%")
+        filter_elem.set("y", "-100%")
+
+        blur_elem = ET.Element("feGaussianBlur")
+        blur_elem.set("in", "SourceGraphic")
+        blur_elem.set("stdDeviation", str(self.contact_distance * 0.4))
+        blur_elem.set("edgeMode", "duplicate")
+
+        filter_elem.append(blur_elem)
+        root.append(filter_elem)
+        color_scale = cm.get_cmap("RdYlGn")
+
+        for key, v in contacts_per_coord.items():
+            x, y = key
+            color = color_scale(1.0 - v * 1.0 / max_contacts)
+            new_elem = ET.Element("circle")
+            new_elem.set("cx", str(x))
+            new_elem.set("cy", str(y))
+            new_elem.set("r", str(self.contact_distance))
+            new_elem.set("fill", color)
+            new_elem.set("fill-opacity", str(0.2))
+            new_elem.set("filter", "url(#blurMe)")
+            root.append(new_elem)
+
+        tree.write(heatmap_file)
