+++ citam/tests/engine/conftest.py
@@ -343,7 +343,6 @@ def simple_facility_model(simple_facility_floorplan, monkeypatch, request):
         occupancy_rate=None,
         buffer=100,
         timestep=1.0,
-        contact_distance=6.0,
         shifts=[{"name": "1", "start_time": 0, "percent_agents": 1.0}],
         meetings_policy_params=None,
         scheduling_policy=None,
